+++ app/src/main/java/com/example/dam_proyecto_final/web_api/WebApiRequest.java
@@ -380,7 +380,8 @@ public class WebApiRequest {
                                     jsonObjectGroup.getInt("groupid"),
                                     jsonObjectGroup.getString("groupname"),
                                     jsonObjectGroup.getString("groupdescription"),
-                                    jsonObjectGroup.getString("currency")
+                                    jsonObjectGroup.getString("currency"),
+                                    jsonObjectGroup.getInt("role")
                             ));
                         }